@@ -3,20 +3,20 @@
   "version": "1.0.0",
   "author": "Neil Bryson Cargamento <neilbrysonmc@gmail.com>",
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "6.2.2"
+    "react": "~18.2.0",
+    "react-dom": "~18.2.0",
+    "react-router-dom": "~6.3.0"
   },
   "devDependencies": {
-    "@neilbryson/eslint-config": "^6.0.1",
+    "@neilbryson/eslint-config": "^8.2.1",
     "@swc/cli": "~0.1.55",
-    "@swc/core": "~1.2.151",
-    "@swc/jest": "~0.2.20",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
-    "@types/react-router-dom": "^5.3.3",
+    "@swc/core": "~1.2.207",
+    "@swc/jest": "~0.2.21",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
     "autoprefixer": "^10.4.2",
     "cross-env": "7.0.3",
     "css-loader": "6.7.1",
@@ -27,18 +27,18 @@
     "fork-ts-checker-webpack-plugin": "^7.2.1",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
-    "jest": "^27.5.1",
+    "jest": "^28.1.1",
     "lint-staged": "^12.3.5",
     "postcss": "8.4.8",
     "postcss-loader": "6.2.1",
     "prettier": "^2.5.1",
     "rimraf": "^3.0.2",
     "style-loader": "3.3.1",
-    "swc-loader": "~0.1.15",
+    "swc-loader": "~0.2.3",
     "tailwindcss": "~3.0.23",
-    "typescript": "~4.6.2",
+    "typescript": "~4.7.4",
     "webpack": "5.70.0",
-    "webpack-cli": "4.9.2",
+    "webpack-cli": "4.10.0",
     "webpack-dev-server": "4.7.4"
   },
   "scripts": {
