@@ -1,6 +1,12 @@
 import './styles/styles.css';
 
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(<React.StrictMode>Hello world</React.StrictMode>, document.getElementById('root'));
+import { RootView } from './views/RootView';
+
+const container = document.getElementById('root');
+
+if (!container) throw new Error('Please add a #root element in public/index.html');
+
+const root = createRoot(container);
+root.render(<RootView />);
